Add metrics and revamp logic
Declare prometheus metrics in gossipsub

Add missing metric labels

Fix metrics types

Add onReportValidation

+++ ts/tracer.ts
@@ -1,8 +1,7 @@
 import { GossipsubIWantFollowupTime } from './constants'
 import { messageIdToString } from './utils'
-import pubsubErrors = require('libp2p-interfaces/src/pubsub/errors')
-
-const { ERR_INVALID_SIGNATURE, ERR_MISSING_SIGNATURE } = pubsubErrors.codes
+import { MsgIdStr, PeerIdStr, RejectReason } from './types'
+import { PromiseDeliveredStats } from './metrics'
 
 /**
  * IWantTracer is an internal tracer that tracks IWANT requests in order to penalize
@@ -17,12 +16,16 @@ export class IWantTracer {
    * Promises to deliver a message
    * Map per message id, per peer, promise expiration time
    */
-  promises = new Map<string, Map<string, number>>()
+  private readonly promises = new Map<MsgIdStr, Map<PeerIdStr, number>>()
+
+  get size(): number {
+    return this.promises.size
+  }
 
   /**
    * Track a promise to deliver a message from a list of msgIds we are requesting
    */
-  addPromise(p: string, msgIds: Uint8Array[]): void {
+  addPromise(from: PeerIdStr, msgIds: Uint8Array[]): void {
     // pick msgId randomly from the list
     const ix = Math.floor(Math.random() * msgIds.length)
     const msgId = msgIds[ix]
@@ -34,17 +37,20 @@ export class IWantTracer {
       this.promises.set(msgIdStr, peers)
     }
 
-    if (!peers.has(p)) {
-      peers.set(p, Date.now() + GossipsubIWantFollowupTime)
+    // If a promise for this message id and peer already exists we don't update the expiry
+    if (!peers.has(from)) {
+      peers.set(from, Date.now() + GossipsubIWantFollowupTime)
     }
   }
 
   /**
    * Returns the number of broken promises for each peer who didn't follow up on an IWANT request.
+   *
+   * This should be called not too often relative to the expire times, since it iterates over the whole data.
    */
-  getBrokenPromises(): Map<string, number> {
+  getBrokenPromises(): Map<PeerIdStr, number> {
     const now = Date.now()
-    const result = new Map<string, number>()
+    const result = new Map<PeerIdStr, number>()
 
     this.promises.forEach((peers, msgId) => {
       peers.forEach((expire, p) => {
@@ -68,18 +74,35 @@ export class IWantTracer {
   /**
    * Someone delivered a message, stop tracking promises for it
    */
-  async deliverMessage(msgIdStr: string): Promise<void> {
+  deliverMessage(msgIdStr: MsgIdStr): PromiseDeliveredStats | null {
+    const expireByPeer = this.promises.get(msgIdStr)
+    if (!expireByPeer) {
+      return null
+    }
+
     this.promises.delete(msgIdStr)
+
+    const now = Date.now()
+    const deliversMs: number[] = []
+
+    for (const expire of expireByPeer.values()) {
+      // time_requested = expire - GossipsubIWantFollowupTime
+      // time_elapsed = now - time_requested
+      deliversMs.push(now - expire - GossipsubIWantFollowupTime)
+    }
+
+    return { requestedCount: expireByPeer.size, deliversMs }
   }
 
   /**
    * A message got rejected, so we can stop tracking promises and let the score penalty apply from invalid message delivery,
    * unless its an obviously invalid message.
    */
-  async rejectMessage(msgIdStr: string, reason: string): Promise<void> {
+  rejectMessage(msgIdStr: MsgIdStr, reason: RejectReason): void {
+    // A message got rejected, so we can stop tracking promises and let the score penalty apply.
+    // With the expection of obvious invalid messages
     switch (reason) {
-      case ERR_INVALID_SIGNATURE:
-      case ERR_MISSING_SIGNATURE:
+      case RejectReason.Error:
         return
     }
